refactor: rename peer_info module to peer_registry
- Update module declaration from peer_info to peer_registry
- Fix all import paths across the codebase to reflect the new name
- Improve naming clarity to better represent the module's purpose
