+++ src/crypto/dkg/joint_feldman/peer_registry.rs
